Fix some more incompatibilities with std::variant from narrowing
std::variant does not allow narrowing conversions. absl::variant didn't
quite get this right, which means narrowing conversions sometimes sneak
in.

The main non-obvious consequence is that variant<size_t, ...> cannot be
constructed from literals like 0 or 2. Such literals have type int, and
an int may be negative. These ints will not be negative, but the
language has lost this information by the time it calls the std::variant
constructor. Writing 0u or 2u works because those have type unsigned,
which can convert to size_t without narrowing. In C++23, we'll have 0uz
and 2uz which even have type size_t, but we're not there yet and no
supported platform has unsigned larger than size_t.

Conversely, a variant<int, ...> cannot be constructed from a size_t
because size_t may not fit in int. That need a checked_cast. There was
also one instance of Persistent<T>, where I used an in_place_type.

Bug: 40242126
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5352774
Commit-Queue: Aya Elsayed <ayaelattar@chromium.org>
Auto-Submit: David Benjamin <davidben@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1270890}

+++ components/enterprise/data_controls/data_controls_policy_handler.cc
@@ -5,6 +5,7 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 #include "components/enterprise/data_controls/data_controls_policy_handler.h"
 
+#include "base/numerics/safe_conversions.h"
 #include "components/enterprise/data_controls/rule.h"
 #include "components/prefs/pref_value_map.h"
 
@@ -54,7 +55,7 @@ bool DataControlsPolicyHandler::CheckPolicySettings(
   for (size_t i = 0; i < rules_list.size(); ++i) {
     DCHECK(rules_list[i].is_dict());
     valid &= Rule::ValidateRuleValue(policy_name(), rules_list[i].GetDict(),
-                                     {i}, errors);
+                                     {base::checked_cast<int>(i)}, errors);
   }
   return valid;
 }
